requirements.txt
# Backend API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Google APIs
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
googlemaps==4.10.0

# Optimization (Using OR-Tools as primary, Gurobi optional)
ortools==9.8.3296
# gurobipy>=11.0.0  # Uncomment if you have Gurobi license

# PDF Generation
reportlab==4.0.8

# Utilities
pydantic==2.5.3
httpx==0.26.0
